Get everything working, begin adding tests for psclient rpc
